tortoolkit/search_server/requirements.txt
aiohttp
fastapi
uvicorn[standard]
beautifulsoup4
aioredis
cloudscraper
fastapi-cache2
lxml
requests
gunicorn